requirements.txt
flask
requests
beautifulsoup4
gunicorn
google-cloud-storage
googlemaps
flask-compress
python-frontmatter
markdown